Move getRemoteHostName and ReverseDNSRunnable into BaseSpnegoKnownClientSystemsFilterAction in order to make functionality common to both derivatives (IP is not normally stored in Active Directory, so queries must be hostname based). Add parameter to query. Drop @NotNull constraint from doExecute function (illegal override; fails at runtime). Add NPE check to LDAP filter in case of 0 results.
